refactor: move page setting to home and thank-you, use property to handle page animate
