+++ src/components/CoverLayout.jsx
@@ -10,96 +10,6 @@ const setTransition = (delaySec) => ({
   duration: 1.0,
 });
 
-const rotateDeg = (rotateTime) => {
-  const finalDeg = -44;
-  return finalDeg + 360 * rotateTime;
-};
-
-const pageSetting = {
-  home: {
-    sZIndex: 'z-10',
-    animate: {
-      letter: {
-        y: [0, -15, 0],
-      },
-      letterC: {
-        initial: {
-          transition: { delayChildren: 2 },
-          transform: 'translate(400px, 10px) rotate(1440deg)',
-        },
-        in: {
-          transform: [
-            `translate(300px, 10px) rotate(${rotateDeg(4)}deg)`,
-            `translate(200px, 10px) rotate(${rotateDeg(3)}deg)`,
-            `translate(100px, 10px) rotate(${rotateDeg(2)}deg)`,
-            `translate(0px, 10px) rotate(${rotateDeg(1)}deg)`,
-            'translate(-82px, 10px) rotate(-54deg)',
-            'translate(-76px, 10px) rotate(-38deg)',
-            'translate(-78px, 10px) rotate(-44deg)',
-          ],
-        },
-      },
-      marker: {
-        rotateZ: 20,
-      },
-    },
-    transition: {
-      letterC: {
-        type: 'tween',
-        ease: 'linear',
-        duration: 4,
-      },
-      marker: {
-        delay: 3.6,
-        duration: 1.0,
-      },
-    },
-    markerIcon: <Icons.HomeTitle.QuestionMark className="z-10" />,
-  },
-  thankyou: {
-    sZIndex: '-z-10',
-    animate: {
-      letter: {
-        rotateZ: [0, 20, 0],
-      },
-      letterC: {
-        initial: {
-          transition: { delayChildren: 2 },
-          transform: 'translate(-66px, 10px) rotate(-44deg)',
-        },
-        in: {
-          transform: [
-            'translate(-82px, 10px) rotate(-44deg)',
-            'translate(-78px, 10px) rotate(25deg)',
-            'translate(-86px, 10px) rotate(-14deg)',
-            'translate(-78px, 10px) rotate(15deg)',
-            'translate(-86px, 10px) rotate(-4deg)',
-            'translate(-78px, 10px) rotate(10deg)',
-            'translate(-82px, 10px) rotate(0deg)',
-          ],
-        },
-      },
-      marker: {
-        y: [0, -40, 0, -30, 0, -10, 0],
-      },
-    },
-    transition: {
-      letterC: {
-        type: 'tween',
-        ease: 'linear',
-        duration: 2,
-      },
-      marker: {
-        delay: 2.6,
-        duration: 1.0,
-      },
-    },
-    markerIcon: <Icons.HomeTitle.ExclamationMark className="z-10" />,
-  },
-};
-
-const setAnimate = (page) => pageSetting[page].animate;
-
 const letterBgArr = [
   {
     class: 'fixed top-[-125px] left-[calc(50%-225px)] -z-10 animate-floating-a',
@@ -118,9 +28,8 @@ const letterBgArr = [
   },
 ];
 
-export function CoverLayout({ info }) {
+export function CoverLayout({ info, pageSetting }) {
   const {
-    page,
     speechBubble,
     bottomText,
     button,
@@ -136,44 +45,44 @@ export function CoverLayout({ info }) {
     {
       id: 0,
       key: 'S',
-      IconHtml: <Icons.HomeTitle.S className={clsx('relative mr-11', pageSetting[page].sZIndex)} />,
+      IconHtml: <Icons.HomeTitle.S className={clsx('relative mr-11', pageSetting.sZIndex)} />,
     },
     {
       id: 1,
       key: 'C',
       initial: 'initial',
       animate: 'in',
-      variants: pageSetting[page].animate.letterC,
-      transition: pageSetting[page].transition.letterC,
+      variants: pageSetting.animate.letterC,
+      transition: pageSetting.transition.letterC,
       className: 'absolute z-0',
       IconHtml: <Icons.HomeTitle.C />,
     },
     {
       id: 2,
       key: 'R',
-      animate: setAnimate(page).letter,
+      animate: pageSetting.animate.letter,
       transition: setTransition(2.0),
       IconHtml: <Icons.HomeTitle.R className="relative -z-10" />,
     },
     {
       id: 3,
       key: 'U',
-      animate: setAnimate(page).letter,
+      animate: pageSetting.animate.letter,
       transition: setTransition(2.2),
       IconHtml: <Icons.HomeTitle.U className="-z-10" />,
     },
     {
       id: 4,
       key: 'M',
-      animate: setAnimate(page).letter,
+      animate: pageSetting.animate.letter,
       transition: setTransition(2.4),
       IconHtml: <Icons.HomeTitle.M className="-z-10" />,
     },
     {
       id: 5,
-      animate: setAnimate(page).marker,
-      transition: pageSetting[page].transition.marker,
-      IconHtml: pageSetting[page].markerIcon,
+      animate: pageSetting.animate.marker,
+      transition: pageSetting.transition.marker,
+      IconHtml: pageSetting.markerIcon,
     },
   ];
